+++ store/src/services/index.ts
@@ -1,1 +1,2 @@
 export * from './permanentStorage/localPermanentStorage.service';
+export * from './temporaryStorage/sessionTemporaryStorage.service';